#23 - Removed ExposedQueryUtils and using Table extensions
